Add files via upload
First commit is the div.v file which is the Xilinx Verilog code for the divider ip_core.  The second commit is the div.xco which is another file attached to the divider ip_core so that it can function.  The third commit is the constraints.ucf file which assigns the inputs and outputs of the counter_program.v file to the pins on the FPGA board.  The final commit is the counter_program.v file which is the Verilog code that allows the divider ip_core to be utilized alongside any other code typed.