Add files via upload
The changes to counter_program includes a clock divider that allows for the counter to count at a frequency of 27 hertz, a reg to keep track of whether or not to engage the divider, a name change for the lower_clk to khz_clk, and another counter and temporary register. The constraints file has the changes of having the lower_clk name changed to khz_clk, and a port dedicated to the hz_clk.